+++ libheimdali/heimdali/regionreader.hxx
@@ -14,13 +14,13 @@ namespace Heimdali {
 class RegionReader
 {
     public:
-        static RegionReader* make_region_reader(size_t sz, size_t sy, size_t nlines_per_loop);
+        static RegionReader* make_region_reader(unsigned int sz, unsigned int sy, unsigned int nlines_per_loop);
         virtual void next_iteration() = 0;
-        virtual void values(size_t &iz, size_t &iy, 
-            size_t &nz, size_t &ny) = 0;
+        virtual void values(unsigned int &iz, unsigned int &iy, 
+            unsigned int &nz, unsigned int &ny) = 0;
         bool is_complete(){return m_is_complete;};
     protected:
-        size_t m_sz,m_sy, m_iz,m_iy, m_nz,m_ny;
+        unsigned int m_sz,m_sy, m_iz,m_iy, m_nz,m_ny;
         bool m_is_complete;
 };
 
@@ -28,22 +28,22 @@ class RegionReaderByLines: public RegionReader
 {
     public:
         RegionReaderByLines(){};
-        RegionReaderByLines(size_t sz, size_t sy, size_t nymax);
+        RegionReaderByLines(unsigned int sz, unsigned int sy, unsigned int nymax);
         void next_iteration();
-        void values(size_t &iz, size_t &iy, size_t &nz, size_t &ny);
+        void values(unsigned int &iz, unsigned int &iy, unsigned int &nz, unsigned int &ny);
     private:
-        size_t m_nymax;
+        unsigned int m_nymax;
 };
 
 class RegionReaderByPlanes: public RegionReader
 {
     public:
         RegionReaderByPlanes(){};
-        RegionReaderByPlanes(size_t sz, size_t sy, size_t nymax);
+        RegionReaderByPlanes(unsigned int sz, unsigned int sy, unsigned int nymax);
         void next_iteration();
-        void values(size_t &iz, size_t &iy, size_t &nz, size_t &ny);
+        void values(unsigned int &iz, unsigned int &iy, unsigned int &nz, unsigned int &ny);
     private:
-        size_t m_nzmax;
+        unsigned int m_nzmax;
 };
 
 };
